# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Unified Authentication Manager API
  description: |
    Production-ready backend API for Unified Authentication Manager.
    
    Features:
    - Zero-knowledge vault (client-side encryption)
    - Device-based MFA with cryptographic signatures
    - Email relay with encrypted routing
    - JWT authentication with refresh tokens
    - Rate limiting and security headers
  version: 1.0.0
  contact:
    name: UAM API Support
    url: https://api.uam.example
servers:
  - url: https://api.uam.example/api/v1
    description: Production server
  - url: http://localhost:5000/api/v1
    description: Local development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
...
    VaultEntry:
      type: object
      required: [ciphertext, nonce, metadata]
      properties:
        ciphertext:
          type: string
          description: AES-256-GCM encrypted blob (base64)
        nonce:
          type: string
        authTag:
          type: string
        metadata:
          type: object
    MfaRequestRequest:
      type: object
      properties:
        deviceId:
          type: string
          format: uuid
        payload:
          type: object
          description: Non-secret metadata about the request
    MfaResponseRequest:
      type: object
      required: [mfaRequestId, signature, decision, timestamp]
      properties:
        mfaRequestId:
          type: string
          format: uuid
        signature:
          type: string
          description: Base64-encoded signature of challenge
        decision:
          type: string
          enum: [approve, deny]
        timestamp:
          type: integer
          format: int64
...
paths:
...
  /vault:
    get:
      tags: [vault]
      summary: List vault entries
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of vault entries
          content:
            application/json:
              schema:
                type: object
                properties:
                  entries:
                    type: array
                    items:
                      $ref: '#/components/schemas/VaultEntry'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [vault]
      summary: Create vault entry
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VaultEntry'
      responses:
        '201':
          description: Entry created
          content:
            application/json:
              schema:
                type: object
                properties:
                  entry:
                    $ref: '#/components/schemas/VaultEntry'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /vault/{id}:
    put:
      tags: [vault]
      summary: Update vault entry
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VaultEntry'
      responses:
        '200':
          description: Entry updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  entry:
                    $ref: '#/components/schemas/VaultEntry'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [vault]
      summary: Delete vault entry
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Entry deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /mfa/request:
    post:
      tags: [mfa]
      summary: Create MFA request
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MfaRequestRequest'
      responses:
        '201':
          description: MFA request created
          content:
            application/json:
              schema:
                type: object
                properties:
                  request:
                    type: object
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /mfa/response:
    post:
      tags: [mfa]
      summary: Submit MFA response with signature
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MfaResponseRequest'
      responses:
        '200':
          description: MFA response processed
          content:
            application/json:
              schema:
                type: object
                properties:
                  request:
                    type: object
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'